+++ backend/handlers/addSubmissionHandler.go
@@ -1,4 +1,4 @@
-package submission
+package handlers
 
 import (
 	"context"
