+++ backend/handlers/changePassword.go
@@ -1,6 +1,7 @@
-package auth
+package handlers
 
 import (
+	"go-test/auth"
 	"go-test/user"
 	"net/http"
 
@@ -24,12 +25,12 @@ func ChangePasswordHandler(c *gin.Context) {
 	// targUser := myTypes.User{UserName: jsonRequest.UserName}
 	// userInfo := user.SearchUser(c, targUser)
 	userInfo := user.GetUserFromUserName(c, jsonRequest.UserName)
-	if !CheckPasswordHash(jsonRequest.ExPass, userInfo.Password) {
+	if !auth.CheckPasswordHash(jsonRequest.ExPass, userInfo.Password) {
 		c.JSON(http.StatusUnauthorized, gin.H{"status": "Password did not match"})
 		return
 	}
 
-	hash, _ := HashPassword(jsonRequest.NewPass)
+	hash, _ := auth.HashPassword(jsonRequest.NewPass)
 	updated := user.UpdateUserPass(c, userInfo.UserName, hash)
 	if updated {
 		c.JSON(http.StatusOK, gin.H{"status": "Password updated successfully!"})
